Refactor types into types directory. Pages get PageContent interfaces
