+++ src/routes/index.tsx
@@ -1,10 +1,16 @@
 import { Resource, component$, Host } from "@builder.io/qwik";
 import { EndpointHandler, useEndpoint } from "@builder.io/qwik-city";
 import Menu from "../components/menu";
-import { MENU, FullMenu, Category } from "../data/menu";
+import { MENU } from "../data/menu";
+import { Category, MenuItem } from "../types";
+
+export interface PageContent {
+  categories: Category[]
+  items: MenuItem[]
+}
 
 export default component$(() => {
-  const menu = useEndpoint<FullMenu>();
+  const menu = useEndpoint<PageContent>();
   return (
     <Host>
       <h2>Welcome to Qwik City Soup &amp; Subs</h2>
@@ -13,7 +19,7 @@ export default component$(() => {
         <h2>Our Menu</h2>
         <Resource
           resource={menu}
-          onResolved={(menu: FullMenu) => (
+          onResolved={(menu) => (
             <Menu categories={menu.categories} items={menu.items} />
           )}
         />
@@ -22,7 +28,7 @@ export default component$(() => {
   );
 });
 
-export const onGet: EndpointHandler<FullMenu> = () => {
+export const onGet: EndpointHandler<PageContent> = () => {
   const categories = MENU.reduce((prev: Category[], { categories }) => {
     return [...prev, ...categories];
   }, []);
